Add a "rewrite announce to create" feature
Adds a checkbox in the Fuzzer called "Rewrite Announce to Create". This is because the Fuzzer only currently supports `Create` activities, but there are lots of good Objects in the Observatory data set that are wrapped in `Announce` due to the Observatory collecting data from relays. If you check this box, any `Announce` activities sent by the Fuzzer will be rewritten as `Create` activities with the same object and should be similar to directly following an account from a service and getting the `Create` instead of the `Announce`.