Fix follows not working
User inbox endpoint and user ID are usually different. If the `object` includes the inbox URI in place of the user id, the follow will fail because there is no followable object there.

This fixes the problem by adding an extra environment variable for the user ID explicitly and using that as the object for a follow activity.